Unnecessary cast removed

+++ src/lib.rs
@@ -155,7 +155,7 @@ impl Playfield {
         let new_y = point.y.checked_add(offset.1);
         if new_x.is_some() && new_y.is_some()
         {
-            { Ok(Point2d {x: (point.x as i64 + offset.0) as i64, y: (point.y as i64 + offset.1) as i64}) }
+            { Ok(Point2d {x: (point.x + offset.0), y: (point.y + offset.1)}) }
         }
         else
         {
